+++ emv_hal.h
@@ -230,8 +230,8 @@ s8 emvHalResetField();
  * bit collision...) occured.
  *****************************************************************************
  */
-s8 emvHalTransceive(const u8 *request, size_t requestLength,
-        u8 *response, size_t maxResponseLength, size_t *responseLength,
+s8 emvHalTransceive(const u8 *request, u32 requestLength,
+        u8 *response, u32 maxResponseLength, u32 *responseLength,
         u32 timeout, EmvHalTransceiveMode_t transceiveMode);
 		
 /*! \ingroup emvHal
