+++ emv_prelayer4.c
@@ -92,8 +92,8 @@
 ******************************************************************************
 */
 
-s8 emvPrelayer4Transceive(const u8 *request, size_t requestLength, u8* response, size_t maxResponseLength,
-    size_t *responseLength, u32 timeout, EmvHalTransceiveMode_t transceiveMode)
+s8 emvPrelayer4Transceive(const u8 *request, u32 requestLength, u8* response, u32 maxResponseLength,
+    u32 *responseLength, u32 timeout, EmvHalTransceiveMode_t transceiveMode)
 {
     int numTries = 0;
